packages/mc-prism/packages/tools/requirements.txt
pytest==7.4.4
black==24.1.1
flake8==7.0.0
mypy==1.8.0
python-dotenv==1.0.0
Pillow>=10.0.0
fonttools>=4.42.0
requests>=2.28.0
numpy>=1.24.0

